fix issue with invalid generated yaml
previous commit to this file removed "default-config:\n" but failed to remove the 5 trailing spaces after this key.  this is causing "personality: IDS\n" to be printed 5 spaces shifted to the right compared to the other keys in the file.  this creates an invalid yaml file and suricata will not load.